Refactor some of the crypto data code
