+++ metro.config.js
@@ -1,0 +1,4 @@
+// Learn more https://docs.expo.io/guides/customizing-metro
+const { getDefaultConfig } = require('expo/metro-config')
+
+module.exports = getDefaultConfig(__dirname)
