+++ package.json
@@ -1,11 +1,10 @@
 {
   "name": "sales-showcase",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
+    "start": "expo start --dev-client",
+    "android": "expo run:android",
+    "ios": "expo run:ios",
     "web": "expo start --web",
     "eject": "expo eject",
     "test": "jest --watchAll",
@@ -35,7 +34,7 @@
     "expo-web-browser": "~10.2.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-native": "0.68.1",
+    "react-native": "0.68.2",
     "react-native-chart-kit": "^6.12.0",
     "react-native-safe-area-context": "4.2.4",
     "react-native-screens": "~3.11.1",
