+++ src/components/partials/CryptoData/index.ts
@@ -1,0 +1,1 @@
+export * from './CryptoData'
