+++ src/components/partials/CryptoData/styled.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components/native'
+
+import { CustomText } from 'common/CustomText'
+
+export const Title = styled(CustomText)`
+  font-size: 16px;
+`
